% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ainm.R
\docType{data}
\name{ainm}
\alias{ainm}
\title{Personal names obtained from CSO data}
\format{A data frame with 5017 rows and 2 variables:
\describe{
 \item{name}{A personal name}
 \item{lexicon}{The source of the word (name), for compatibility with tidytext}
 }}
\usage{
ainm
}
\description{
Personal names with 3 occurences or more, from CSO via data.gov.ie
}
\keyword{datasets}
