% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_with_lens.R
\name{bivariate_split}
\alias{bivariate_split}
\title{Bivariate split}
\usage{
bivariate_split(split, x = "from_x", y = "from_y", metric = NULL)
}
\arguments{
\item{split}{data.frame with the connections (categories) of two-dimensional lenses}

\item{x}{string variable name for the first dimension of the lens. Default 'from_x'.}

\item{y}{string variable name for the second dimension of the lens. Default 'from_y'.}

\item{metric}{string or array defining the metrics supported ("polar" or "euclidean").}
}
\value{
Returns a \code{data.frame} with the preserved connections.
}
\description{
Filter values connections for two-dimensional lens
}
