% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circles}
\alias{circles}
\title{Circles}
\format{A data frame of 7 rows and 3 columns
\describe{
\item{x}{Numeric values for X dimension}
\item{y}{Numeric value for Y dimension}
\item{lens}{Numeric value used to color the nodes}
}}
\source{
Manually generated data
}
\usage{
circles
}
\description{
A dataset generated to illustrate STAD functionality.
}
\keyword{datasets}
