% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_with_lens.R
\name{custom_distance}
\alias{custom_distance}
\title{Custom distance}
\usage{
custom_distance(x, metric)
}
\arguments{
\item{x}{array variable. Dimension of the lens.}

\item{metric}{string or array defining the metrics supported ("polar" or "euclidean").}
}
\value{
Returns a \code{dist} object with the distance of the filter.
}
\description{
Internal distance matrix for bivariate_split. Uses polar or euclidean metric. Returns a distance matrix as sum of the two independents dimensions.
}
