% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_without_lens.R
\name{generate_graph}
\alias{generate_graph}
\title{Generate graph}
\usage{
generate_graph(ordered_distances = NULL, number_edges = NULL,
  vertices_names = NULL)
}
\arguments{
\item{ordered_distances}{data.frame with the information from distance matrix and MST.}

\item{number_edges}{numeric value that represents the number of edges needed to generate the graph.}

\item{vertices_names}{numeric array with names of nodes (\code{colnames(distance_matrix)})}
}
\value{
Returns a \code{igraph} object with the defined graph.
}
\description{
Internal function to generate graph from ordered_distances data.frame using the number of edges
}
