% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_without_lens.R
\name{modify_distance_matrix}
\alias{modify_distance_matrix}
\title{Modify distance matrix}
\usage{
modify_distance_matrix(distance_matrix)
}
\arguments{
\item{distance_matrix}{dist object containing a distance matrix.}
}
\value{
Returns a \code{list} with the following items:
\itemize{
\item distance_matrix_df
\item vertices_names
}
}
\description{
Internal function that transforms the distance matrix into a dataframe. Additionaly, this function provides the array vertices_names for the graph generation.
}
