% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_graph.R
\name{modify_graph}
\alias{modify_graph}
\title{Modify graph}
\usage{
modify_graph(x, number_edges)
}
\arguments{
\item{x}{\code{stad} object}

\item{number_edges}{string variable name for the first dimension of the lens. Default 'from_x'.}
}
\value{
Returns a \code{stad} class (or \code{list}) with the following items:
\itemize{
\item graph
\item number_edges
\item trace
\item ordered_distances
}
}
\description{
Update a STAD graph with the a new number of edges.
}
