% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_without_lens.R
\name{modify_output}
\alias{modify_output}
\title{Modify output from optim_sa}
\usage{
modify_output(output)
}
\arguments{
\item{output}{String with the intermediate results from optim_sa.}
}
\value{
Returns a \code{list} object with following items:
\itemize{
\item max_value
\item trace
}
}
\description{
Transform the output from the intermediate results and save it as a \code{data.frame}.
}
