% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot graph}
\usage{
plot_graph(x, layout = igraph::layout_nicely, vertex.color = "black",
  vertex.frame.color = "grey", vertex.size = 3, edge.width = 0.5)
}
\arguments{
\item{x}{\code{stad} object.}

\item{layout}{igraph layout. Default igraph::layout_nicely}

\item{vertex.color}{Vertex color for the plot. Default black.}

\item{vertex.frame.color}{Vertex color contour. Default grey.}

\item{vertex.size}{Vertex size. Default 3}

\item{edge.width}{Edge width. Default 0.5}
}
\description{
Graph plot for the resulting STAD graph.
}
\examples{
# Iris dataset
data(iris)
iris_distance <- dist(iris[,1:4])
iris_stad <- stad(iris_distance)
plot_graph(iris_stad)
plot_trace(iris_stad)

}
