% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot trace}
\usage{
plot_trace(x)
}
\arguments{
\item{x}{\code{stad} object.}
}
\description{
Line-chart of the intermediate evaluations during the optimizations process.
}
\examples{
# Iris dataset
data(iris)
iris_distance <- dist(iris[,1:4])
iris_stad <- stad(iris_distance)
plot_graph(iris_stad)
plot_trace(iris_stad)

}
