% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_without_lens.R
\name{stad_evaluation}
\alias{stad_evaluation}
\title{STAD evaluation}
\usage{
stad_evaluation(distance_matrix = NULL, distance_matrix_df = NULL,
  vertices_names = NULL, mst_graph = NULL, penalty = NULL,
  random_factor = NULL, iterations_inner_loop = NULL)
}
\arguments{
\item{distance_matrix}{Distance matrix as \code{dist} format.}

\item{distance_matrix_df}{Distance matrix as \code{data.frame} format.}

\item{vertices_names}{numeric array with names of nodes (\code{colnames(distance_matrix)})}

\item{mst_graph}{Minimum Spanning Tree graph from distance_matrix}

\item{penalty}{numeric penalty value used during the optimization loop. It limits the number of iterations. Default = 0.}

\item{random_factor}{integer factor that controls the variability of the next iteration on the optimization algorithm. Default = 10000. Higher values generates more diverse exploration.}

\item{iterations_inner_loop}{integer number of evaluations for each iteration. The number of iterations allows knowing the close values around but increase the total number of iterations.}
}
\value{
Returns a \code{stad} class (or \code{list}) with the following items:
\itemize{
\item graph
\item number_edges
\item trace
\item ordered_distances
}
}
\description{
Iterative process in STAD using Simulated Annealing algorithm
}
