% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_without_lens.R
\name{stad_without_lens}
\alias{stad_without_lens}
\title{STAD without lens}
\usage{
stad_without_lens(distance_matrix, penalty = 0, random_factor = 10000,
  iterations_inner_loop = 10)
}
\arguments{
\item{distance_matrix}{dist object containing a distance matrix.}

\item{penalty}{numeric penalty value used during the optimization loop. It limits the number of iterations. Default = 0.}

\item{random_factor}{integer factor that controls the variability of the next iteration on the optimization algorithm. Default = 10000. Higher values generates more diverse exploration.}

\item{iterations_inner_loop}{integer number of evaluations for each iteration. The number of iterations allows knowing the close values around but increase the total number of iterations.}
}
\value{
Returns a \code{stad} class (or \code{list}) with the following items:
\itemize{
\item graph
\item number_edges
\item trace
\item ordered_distances
}
}
\description{
The function computes STAD without lens using a given distance matrix.
}
