% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuairim.R
\docType{data}
\name{tuairim}
\alias{tuairim}
\title{Sentiment words for Irish}
\format{A data frame with 963 rows and 2 variables:
\describe{
 \item{word}{A word}
 \item{sentiment}{The sentiment of the word, positive or negative}
 }}
\source{
\itemize{
 \item \url{https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
 \item \url{https://github.com/juliasilge/tidytext/tree/master/data-raw}
 }
}
\usage{
tuairim
}
\description{
Sentiment words for Irish, a loose translation from Bing Liu lexicon
in tidytext
}
\keyword{datasets}
