% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_with_lens.R
\name{univariate_intra}
\alias{univariate_intra}
\title{Univariate intra-edge}
\usage{
univariate_intra(from_x, to_x, range, metric)
}
\arguments{
\item{from_x}{numeric; categroy of the group (interval) for from element in the edge.}

\item{to_x}{numeric; categroy of the group (interval) for to element in the edge.}

\item{range}{array. Numeric range of the lens}

\item{metric}{string or array defining the metrics supported ("polar" or "euclidean").}
}
\value{
Returns a logical element.
}
\description{
Definition of intra-edge for one-dimensional lens
}
