% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad.R
\docType{data}
\name{stad}
\alias{stad}
\title{Stop words for Irish}
\format{A data frame with 163 rows and 2 variables:
\describe{
 \item{word}{An Irish word}
 \item{lexicon}{The source of the word, for compatibility with tidytext}
 }}
\usage{
stad
}
\description{
Stop words for Irish
}
\keyword{datasets}
